Rename Arball to Orbit in Unit Tests
There's really not much more to say here.
